discord-frame-bot-with-listframes/requirements.txt
discord.py>=2.0.0
Pillow
requests
python-dotenv